Update 100-atoi.c comit

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -9,7 +9,7 @@
 int _atoi(char *s)
 {
 	char sign;
-	long long int n;
+	int n;
 
 	sign = '+';
 	n = 0;
